refactor: extract getExpectedFieldData, getStyleOfAnElement and getChildrenDirection into a separate utils file
