+++ src/liveEditor/generators/generatePseudoEditableField.tsx
@@ -1,68 +1,5 @@
-import { toString } from "lodash-es";
-import { CslpData } from "../../cslp/types/cslp.types";
-import liveEditorPostMessage from "../utils/liveEditorPostMessage";
-import { LiveEditorPostMessageEvents } from "../utils/types/postMessage.types";
 import { render } from "preact";
 import PseudoEditableFieldComponent from "../components/pseudoEditableField";
-
-/**
- * Retrieves the expected field data based on the provided field metadata.
- *
- * @param fieldMetadata The metadata of the field.
- * @returns A promise that resolves to the expected field data as a string.
- */
-export async function getExpectedFieldData(
-    fieldMetadata: CslpData
-): Promise<string> {
-    const data = await liveEditorPostMessage?.send<{ fieldData: unknown }>(
-        LiveEditorPostMessageEvents.GET_FIELD_DATA,
-        { fieldMetadata }
-    );
-
-    return toString(data?.fieldData);
-}
-
-/**
- * Retrieves the computed style of an HTML element.
- *
- * @param element - The HTML element to retrieve the style from.
- * @returns An object representing the computed style of the element.
- */
-export function getStyleOfAnElement(element: HTMLElement): {
-    [key: string]: string;
-} {
-    const styleSheetDeclaration = window.getComputedStyle(element);
-    const styleSheetArray = Array.from(styleSheetDeclaration);
-
-    const FILTER_STYLES = [
-        "position",
-        "left",
-        "top",
-        "right",
-        "bottom",
-        "text-overflow",
-        "margin",
-        "margin-block-end",
-        "margin-block-start",
-        "margin-inline-end",
-        "margin-inline-start",
-        "margin-left",
-        "margin-right",
-        "margin-top",
-        "margin-bottom",
-        "-webkit-user-modify",
-    ];
-
-    const styles: { [key: string]: string } = {};
-    styleSheetArray.forEach((style: string) => {
-        if (!FILTER_STYLES.includes(style)) {
-            const styleValue = styleSheetDeclaration.getPropertyValue(style);
-            styles[style] = styleValue;
-        }
-    });
-
-    return styles;
-}
 
 /**
  * Checks if the content of an element is truncated with ellipsis.
